+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,17 +1,12 @@
+package blackjack.domain;
+
 import java.util.List;
 
-public class Player {
-    private final String name;
-
+public class Dealer {
     private Cards drawnCards;
 
-    public Player(String name) {
-        this.name = name;
-        this.drawnCards = new Cards();
-    }
-
-    public String getName() {
-        return name;
+    public Dealer() {
+        drawnCards = new Cards();
     }
 
     public void receiveOneCard(Card card) {
@@ -26,11 +21,19 @@ public class Player {
         return drawnCards.getCardNames();
     }
 
+    public int getAdditionalDrawnCardCount() {
+        return drawnCards.getCardNames().size() - 2;
+    }
+
     public boolean isBust() {
         return drawnCards.sumScore() > 21;
     }
 
     public int calculateTotalScore() {
         return drawnCards.sumScore();
+    }
+
+    public boolean isDealerDone() {
+        return calculateTotalScore() >= 17;
     }
 }
