+++ src/main/java/blackjack/domain/GameResult.java
@@ -1,3 +1,5 @@
+package blackjack.domain;
+
 import java.util.Map;
 
 public record GameResult(
