+++ src/main/java/blackjack/domain/Player.java
@@ -1,11 +1,19 @@
-import java.util.ArrayList;
+package blackjack.domain;
+
 import java.util.List;
 
-public class Dealer {
+public class Player {
+    private final String name;
+
     private Cards drawnCards;
 
-    public Dealer() {
-        drawnCards = new Cards();
+    public Player(String name) {
+        this.name = name;
+        this.drawnCards = new Cards();
+    }
+
+    public String getName() {
+        return name;
     }
 
     public void receiveOneCard(Card card) {
@@ -20,19 +28,11 @@ public class Dealer {
         return drawnCards.getCardNames();
     }
 
-    public int getAdditionalDrawnCardCount() {
-        return drawnCards.getCardNames().size() - 2;
-    }
-
     public boolean isBust() {
         return drawnCards.sumScore() > 21;
     }
 
     public int calculateTotalScore() {
         return drawnCards.sumScore();
-    }
-
-    public boolean isDealerDone() {
-        return calculateTotalScore() >= 17;
     }
 }
