+++ src/main/java/blackjack/domain/ScoreCompareResult.java
@@ -1,3 +1,5 @@
+package blackjack.domain;
+
 public enum ScoreCompareResult {
     PLAYER_WIN,
     DEALER_WIN,
