+++ src/main/java/blackjack/view/InputView.java
@@ -1,3 +1,5 @@
+package blackjack.view;
+
 import java.util.Scanner;
 
 public class InputView {
