+++ src/test/java/DealerTest.java
@@ -1,5 +1,7 @@
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.domain.Card;
+import blackjack.domain.Dealer;
 import org.junit.jupiter.api.Test;
 
 public class DealerTest {
